Add files via upload
This is a simple Java program that lets users log in by entering a username and password. The password input is hidden using asterisks for security. It gives the user three tries to get the correct credentials and shows messages based on whether the login is successful or not. The code is straightforward and includes comments to make it easy to follow.